* Clarified /delayKill in plugin.yml * Fixed immutable array issue in TC++ language database * Fixed issue with no search = no commands in /troll * Added data ability to GuiRegistry

+++ src/me/egg82/tcpp/services/GuiRegistry.java
@@ -13,7 +13,7 @@ public class GuiRegistry extends Registry {
 		
 		setRegister("alone", Material.class, Material.IRON_DOOR);
 		setRegister("annoy", Material.class, Material.MONSTER_EGG);
-		setRegister("annoy_data", Short.class, 120);
+		setRegister("annoy_data", Short.class, (short) 120);
 	}
 	
 	//public
